campaign_spend_insight/requirements.txt
google-adk[google-genai]
pandas
